+++ eShop.Messaging/Contracts/ProductPrice.cs
@@ -3,7 +3,7 @@
     public class ProductPrice
     {
         public Currency Currency { get; set; }
-        public double Price { get; set; }
-        public double? DiscountedPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal? DiscountedPrice { get; set; }
     }
 }
