+++ eShop.Viber/ViberBotFramework/Controllers/ComissionSettingsController.cs
@@ -72,7 +72,7 @@ namespace eShop.Viber.ViberBotFramework.Controllers
             var user = await _viberService.GetUserByIdAsync(context.UserId);
             if (user!.AccountId != null)
             {
-                if (double.TryParse(context.Text, out var amount))
+                if (decimal.TryParse(context.Text, out var amount))
                 {
                     var request = new SetComissionAmountRequest(user.AccountId.Value, amount);
                     var response = await _setComissionAmounRequestClient.GetResponse<SetComissionAmountResponse>(request);
